fix: Adapt codebase to changes in sequoia-sq 0.32.0
The behavior of `sq packet split` has changed and we are dropping the
`--prefix` option, as it no longer allows for providing an empty string.
Move the filename endings of all understood OpenPGP packet types
resulting from `sq packet split` calls to a unique Enum, so it is easier
to refer to the filenames.

Signed-off-by: David Runge <dvzrv@archlinux.org>